pyproject.toml
[project]
authors = [
    {name = "Yurasov Aleksey", github = "https://github.com/YurasovAleksey"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.pytest.ini_options]
addopts = "--cov=gendiff --cov-report=xml"
